requirements.txt
Adafruit_DHT >= 1.4.0
python-dotenv >= 0.21.1
paho-mqtt >= 2.1.0
